requirements.txt
streamlit
PyMuPDF
faiss-cpu
sentence-transformers
requests
